Modified get_atmosphere and get_density when single float is input for speed.
The present functions utilizing np.where (which works for single floats and arrays) are quite inefficient in case of an float. Since these function are in the 'inner loop' of most calculations and mostly getting floats there is a benefit in having this duplicated code pieces.
For get_distance_xmax_geometrical I got a boost of an factor of 3 in computing time.